+++ sys/platform/pc64/isa/prof_machdep.c
@@ -25,7 +25,6 @@
  * SUCH DAMAGE.
  *
  * $FreeBSD: src/sys/i386/isa/prof_machdep.c,v 1.14.2.1 2000/08/03 00:09:30 ps Exp $
- * $DragonFly: src/sys/platform/pc64/isa/prof_machdep.c,v 1.1 2008/08/29 17:07:20 dillon Exp $
  */
 
 #ifdef GUPROF
@@ -235,7 +234,7 @@ cputime(void)
 	 */
 	delta = prev_count - count;
 	prev_count = count;
-	if ((int) delta <= 0)
+	if (delta <= 0)
 		return (delta + (timer0_max_count << CPUTIME_CLOCK_I8254_SHIFT));
 	return (delta);
 }
